+++ Core/Dialog/PSPMsgDialog.cpp
@@ -186,7 +186,7 @@ void PSPMsgDialog::DisplayMessage(std::string text, bool hasYesNo, bool hasOK) {
 
 	// Without the scrollbar, we have 390 total pixels.
 	float WRAP_WIDTH = 340.0f;
-	if (UTF8StringNonASCIICount(text.c_str()) >= text.size() / 4) {
+	if ((size_t)UTF8StringNonASCIICount(text.c_str()) >= text.size() / 4) {
 		WRAP_WIDTH = 376.0f;
 		if (text.size() > 12) {
 			messageStyle.scale = 0.6f;
